Create interface SearchView and move Similar auctions in SearchView
